#62. Save problems to database using Room.

+++ app/src/main/java/com/bogdan/codeforceswatcher/room/RoomController.kt
@@ -1,6 +1,7 @@
 package com.bogdan.codeforceswatcher.room
 
 import com.bogdan.codeforceswatcher.features.contests.redux.states.ContestsState
+import com.bogdan.codeforceswatcher.features.problems.redux.states.ProblemsState
 import com.bogdan.codeforceswatcher.features.users.redux.states.UsersState
 import com.bogdan.codeforceswatcher.redux.states.AppState
 import com.bogdan.codeforceswatcher.store
@@ -13,6 +14,7 @@ object RoomController : StoreSubscriber<AppState> {
         store.subscribe(this) {
             it.skipRepeats { oldState, newState ->
                 oldState.contests == newState.contests
+                    && oldState.problems == newState.problems
             }
         }
     }
@@ -21,15 +23,23 @@ object RoomController : StoreSubscriber<AppState> {
         return AppState(
             contests = ContestsState(contests = DatabaseClient.contestDao.getUpcomingContests().reversed()),
             users = UsersState(
-                users = DatabaseClient.userDao.getAll(),
+                users = DatabaseClient.userDao.getUsers(),
                 sortType = UsersState.SortType.getSortType(Prefs.get().readSpinnerSortPosition().toInt())
+            ),
+            problems = ProblemsState(
+                problems = DatabaseClient.problemsDao.getProblems()
             )
         )
     }
 
     override fun newState(state: AppState) {
-        DatabaseClient.contestDao.deleteAll()
-        DatabaseClient.contestDao.insert(state.contests.contests)
-        // TODO add updating db with problems
+        if (DatabaseClient.contestDao.getUpcomingContests() != state.contests.contests) {
+            DatabaseClient.contestDao.deleteAll()
+            DatabaseClient.contestDao.insert(state.contests.contests)
+        }
+        if (DatabaseClient.problemsDao.getProblems() != state.problems.problems) {
+            DatabaseClient.problemsDao.deleteAll()
+            DatabaseClient.problemsDao.insert(state.problems.problems)
+        }
     }
 }
